Always remove oldest pins first
If we load so much data that we run over our MAX_PUSHPIN_COUNT, then we
need to start removing pins. Previously, we would remove the last items
added to the list. However, these items were not necessarily the
pushpins with the oldest dates. This change updates the crime binding to
always be a sorted projection of the underlying list binding so that
pushes and pops are always in date sorted order.